Buttons Navegation Done
Implement a button navegation in a separete scipt at the MainManager object, it works and I could use the same code of the Back Menu in the same function button. To it works I needed to creat a prefab of the MainManager object and use it to attach to the buttons in every Scene.